+++ src/config.ts
@@ -1,0 +1,2 @@
+export const djinnjsOutDir = "REPLACE_WITH_OUTPUT_DIR_NAME";
+export const minimumConnection = "REPLACE_WITH_MINIMUM_CONNECTION";